# docs/static/assets/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactSVGUploadTO": {
        "type": "object",
        "required": [
          "svgPreview"
        ],
        "properties": {
          "svgPreview": {
            "type": "string"
          }
        }
      },
      "ArtifactUpdateTO": {
        "type": "object",
        "required": [
          "description",
          "name"
        ],
        "properties": {
          "svgPreview": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          }
        }
      },
      "ArtifactVersionUploadTO": {
        "type": "object",
        "properties": {
          "saveType": {
            "type": "string",
            "enum": [
              "MILESTONE",
              "AUTOSAVE"
            ]
          },
          "xml": {
            "type": "string"
          },
          "versionComment": {
            "type": "string"
          }
        }
      },
...
      "ArtifactTO": {
        "type": "object",
        "required": [
          "description",
          "name",
          "repositoryId"
        ],
        "properties": {
          "createdDate": {
            "format": "date-time",
            "type": "string"
          },
          "svgPreview": {
            "type": "string"
          },
          "repositoryId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string"
          },
          "fileType": {
            "type": "string"
          }
        }
      },
...
      "NewArtifactTO": {
        "type": "object",
        "required": [
          "description",
          "name"
        ],
        "properties": {
          "svgPreview": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
...
    "/api/artifact/{repositoryId}": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewArtifactTO"
              }
            }
          },
          "required": true
        },
        "operationId": "createArtifact",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactTO"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repositoryId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
...
    "/api/artifact/previewSVG/{artifactId}": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactSVGUploadTO"
              }
            }
          },
          "required": true
        },
        "operationId": "updatePreviewSVG",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
...
    "/api/artifact/{artifactId}/unlock": {
      "post": {
        "operationId": "unlockArtifact",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
    "/api/artifact/{artifactId}": {
      "get": {
        "operationId": "getArtifact",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactTO"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      },
      "delete": {
        "summary": "Delete one Artifact and all of its versions",
        "operationId": "deleteArtifact",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactUpdateTO"
              }
            }
          },
          "required": true
        },
        "operationId": "updateArtifact",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactTO"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
...
    "/api/artifact/search/{typedTitle}": {
      "get": {
        "operationId": "searchArtifacts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactTO"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "typedTitle",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
...
    "/api/version/{artifactId}": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactVersionUploadTO"
              }
            }
          },
          "required": true
        },
        "operationId": "createOrUpdateVersion",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Version"
        ]
      }
    },
...
    "/api/artifact/{artifactId}/lock": {
      "post": {
        "operationId": "lockArtifact",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
    "/api/artifact/repository/{repositoryId}": {
      "get": {
        "operationId": "getArtifactsFromRepo",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactTO"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repositoryId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
...
    "/api/artifact/starred/{artifactId}": {
      "post": {
        "operationId": "setStarred",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "artifactId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  }
}